Contemplating killing myself if this doesn't work

+++ backend/src/main/java/com/eman/service/EmployeeService.java
@@ -1,7 +1,6 @@
 package com.eman.service;
 
 import java.util.List;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import com.eman.entity.Employee;
@@ -17,15 +16,16 @@ public class EmployeeService {
         return employeeRepository.findAll();
     }
 
-    public Employee getEmployeeById(Long id) {
-        return employeeRepository.findById(id).orElseThrow(() -> new RuntimeException("Employee not found"));
+    public Employee getEmployeeById(Integer id) {  // Changed from Long to Integer
+        return employeeRepository.findById(id)
+            .orElseThrow(() -> new RuntimeException("Employee not found"));
     }
 
     public Employee saveEmployee(Employee employee) {
         return employeeRepository.save(employee);
     }
 
-    public void deleteEmployee(Long id) {
+    public void deleteEmployee(Integer id) {  // Changed from Long to Integer
         employeeRepository.deleteById(id);
     }
 }
